+++ scripts/push-activity-js-runtime.sh
@@ -21,8 +21,9 @@ OUTPUT_FILE="${2:-crates/embedded-assets/activity-js-runtime-version.txt}"
 cargo check --package activity-js-runtime-builder # triggers build.rs of activity-js-runtime-builder
 
 if [ "$TAG" != "dry-run" ]; then
-    STRIPPED="target/release_wasm_runtime/wasm32-wasip2/release_wasm_runtime/activity_js_runtime.stripped.wasm"
-    wasm-tools strip --all "target/release_wasm_runtime/wasm32-wasip2/release_wasm_runtime/activity_js_runtime.wasm" -o "$STRIPPED"
+    CARGO_TARGET_ROOT=${CARGO_TARGET_DIR:-target}
+    STRIPPED="$CARGO_TARGET_ROOT/release_wasm_runtime/wasm32-wasip2/release_wasm_runtime/activity_js_runtime.stripped.wasm"
+    wasm-tools strip --all "$CARGO_TARGET_ROOT/release_wasm_runtime/wasm32-wasip2/release_wasm_runtime/activity_js_runtime.wasm" -o "$STRIPPED"
     TMP_TOML="activity-deployment-for-push.toml"
     trap "rm -f $TMP_TOML" EXIT
     cat > "$TMP_TOML" <<EOF